+++ changelog.md
@@ -1,4 +1,9 @@
 # 1.2.0
+- Add __Zoink's__ completion of [Orbit](https://youtu.be/QKcv8DkNPd0?si=OJdK0ou-j5u4SI0g)
+- Add __Kingsammelot's__ completion of [Artificial Ascent](https://www.youtube.com/watch?v=q8f7yzvM8EQ)
+- Add __Npesta's__ completion of [Deimos (EndLevel)](https://www.youtube.com/watch?v=FgxsNzQnF5o)
+- Add __AeonAir's__ completion of [Tartarus](https://www.youtube.com/watch?v=areGj_ppaYo)
+- Add __Npesta's__ completion of [WOW](https://www.youtube.com/watch?v=CbUVw7p0Ebg)
 - Add ability to use a specific reaction instead of a random one
 - Normalized reaction volume and made them louder (plz turn down your volume)
 - Fix reactions not working with end triggers